+++ MANIFEST.in
@@ -1,3 +1,2 @@
 include casparser_isin/*.py
-include casparser_isin/VERSION.txt
 include casparser_isin/isin.db
