--- casparser_isin/VERSION.txt
@@ -1,1 +1,0 @@
-2023.8.18
