+++ casparser_isin/__init__.py
@@ -1,7 +1,8 @@
 from .mf_isin import MFISINDb
-from .__version__ import __version__
 
 __all__ = [
     "MFISINDb",
     "__version__",
 ]
+
+__version__ = "2023.9.3"
